# Speed-up normalization layers