- Speed-up other normalization layers, by re-arranging expression a bit.

```math
y = \frac{x - E[x]}{\sqrt{Var[x] + \epsilon}} * \gamma + \beta
  = \frac{\gamma}{\sqrt{Var[x] + \epsilon}} * (x - E[x]) + \beta
```
```math
scale = \frac{\gamma}{\sqrt{Var[x] + \epsilon}}
```
```math
bias = -scale * E[x] + \beta
```
```math
y = scale * x + bias
```

```julia
julia> using BenchmarkTools, Flux

julia> gn = GroupNorm(128, 32);

julia> x = rand(Float32, 256, 256, 128, 16);

julia> @benchmark gn(x)
```

**Before:**

```julia
julia> @benchmark gn(x)
BenchmarkTools.Trial: 9 samples with 1 evaluation.
 Range (min … max):  545.960 ms … 576.558 ms  ┊ GC (min … max): 0.27% … 6.13%
 Time  (median):     575.436 ms               ┊ GC (median):    6.13%
 Time  (mean ± σ):   570.931 ms ±  10.306 ms  ┊ GC (mean ± σ):  5.27% ± 2.01%

                                                              █  
  ▇▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▇▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▇▁▇▇▇█ ▁
  546 ms           Histogram: frequency by time          577 ms <

 Memory estimate: 1.00 GiB, allocs estimate: 32.
```

**This PR:**

```julia
julia> @benchmark gn(x)
BenchmarkTools.Trial: 23 samples with 1 evaluation.
 Range (min … max):  202.554 ms … 221.089 ms  ┊ GC (min … max): 0.34% … 7.94%
 Time  (median):     219.391 ms               ┊ GC (median):    7.99%
 Time  (mean ± σ):   218.166 ms ±   4.402 ms  ┊ GC (mean ± σ):  7.34% ± 2.20%

                                                        ▆█▂      
  ▄▁▁▁▁▁▁▁▁▁▁▁▄▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▄▄███▄▁▁▄ ▁
  203 ms           Histogram: frequency by time          221 ms <

 Memory estimate: 512.03 MiB, allocs estimate: 35.
```

### PR Checklist

- [ ] Tests are added
- [ ] Entry in NEWS.md
- [ ] Documentation, if applicable