@pxl-th 2023-03-28
**Before:**

```julia
julia> @benchmark Zygote.gradient(gn -> sum(gn(x)), gn)
BenchmarkTools.Trial: 4 samples with 1 evaluation.
 Range (min … max):  1.530 s …   1.593 s  ┊ GC (min … max): 3.85% … 4.30%
 Time  (median):     1.570 s              ┊ GC (median):    3.90%
 Time  (mean ± σ):   1.566 s ± 31.406 ms  ┊ GC (mean ± σ):  3.83% ± 0.54%

  █                █                                    █ █  
  █▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁█▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁█▁█ ▁
  1.53 s         Histogram: frequency by time        1.59 s <

 Memory estimate: 5.50 GiB, allocs estimate: 611.
```

**This PR:**

```julia

julia> @benchmark Zygote.gradient(gn -> sum(gn(x)), gn)
BenchmarkTools.Trial: 5 samples with 1 evaluation.
 Range (min … max):  1.026 s …   1.057 s  ┊ GC (min … max): 4.64% … 5.15%
 Time  (median):     1.040 s              ┊ GC (median):    4.58%
 Time  (mean ± σ):   1.040 s ± 14.209 ms  ┊ GC (mean ± σ):  4.38% ± 0.96%

  █                        ▁                      ▁       ▁  
  █▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁█▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁█▁▁▁▁▁▁▁█ ▁
  1.03 s         Histogram: frequency by time        1.06 s <

 Memory estimate: 3.50 GiB, allocs estimate: 675.
```