@chengchingwen 2023-03-28
And a potential regression issue. The result is now aligned to the pytorch norm, but that means all old model trained with old Flux norm will be affected.