@pxl-th 2023-03-28
> What about benchmark on gpu?

**Before:**

```julia
julia> @benchmark CUDA.@sync gn(x)
BenchmarkTools.Trial: 109 samples with 1 evaluation.
 Range (min … max):  41.979 ms … 52.871 ms  ┊ GC (min … max): 0.00% … 6.65%
 Time  (median):     44.962 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   45.950 ms ±  3.069 ms  ┊ GC (mean ± σ):  1.51% ± 2.50%

    ▂         ▁▂  █ ▁                                          
  ▃▆█▄█▄▆▆▁▁▃▄███▇█▆█▇▁▃▁▁▁▁▁▁▃▁▃▁▁▁▁▁▁▃▄▁▄█▃▃▃▃▃▄▄▇▁▄▄▃▁▁▁▁▃ ▃
  42 ms           Histogram: frequency by time        52.8 ms <

 Memory estimate: 28.30 KiB, allocs estimate: 321.


julia> @benchmark CUDA.@sync Zygote.gradient(gn -> sum(gn(x)), gn)
BenchmarkTools.Trial: 25 samples with 1 evaluation.
 Range (min … max):  200.516 ms … 210.522 ms  ┊ GC (min … max): 4.68% … 1.57%
 Time  (median):     202.030 ms               ┊ GC (median):    4.64%
 Time  (mean ± σ):   202.604 ms ±   2.110 ms  ┊ GC (mean ± σ):  4.49% ± 0.63%

      ▁   █ ▁▄   ▁                                               
  ▆▆▁▆█▆▆▆█▆██▁▁▆█▁▁▁▆▆▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▆▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▆ ▁
  201 ms           Histogram: frequency by time          211 ms <

 Memory estimate: 93.84 KiB, allocs estimate: 1551.
```

**This PR:**

```julia
julia> @benchmark CUDA.@sync gn(x)
BenchmarkTools.Trial: 153 samples with 1 evaluation.
 Range (min … max):  30.696 ms … 39.091 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     32.063 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   32.806 ms ±  2.100 ms  ┊ GC (mean ± σ):  0.88% ± 2.26%

          ▁█▄                                                  
  ▄▃▃▃▃▄▅▇███▅▆▅▄▁▃▃▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▂▂▁▄▂▂▁▂▁▂▂▂▃▂▃▂▂▂ ▂
  30.7 ms         Histogram: frequency by time        38.8 ms <

 Memory estimate: 34.22 KiB, allocs estimate: 385.

julia> @benchmark CUDA.@sync Zygote.gradient(gn -> sum(gn(x)), gn)
BenchmarkTools.Trial: 39 samples with 1 evaluation.
 Range (min … max):  126.358 ms … 138.385 ms  ┊ GC (min … max): 2.34% … 2.65%
 Time  (median):     130.110 ms               ┊ GC (median):    4.00%
 Time  (mean ± σ):   130.436 ms ±   2.064 ms  ┊ GC (mean ± σ):  3.75% ± 0.89%

                   █▅▂ ▅▂                                        
  ▅▁▁▅▁▁▁▅▁▁▁▅▁█▁█████▅██▅▁█▅▅▅▁▁▁▁▅▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▅▁▁▁▁▁▁▁▁▁▅ ▁
  126 ms           Histogram: frequency by time          138 ms <

 Memory estimate: 99.58 KiB, allocs estimate: 1647.
```