@CarloLucibello 2023-03-29
> And a potential regression issue. The result is now aligned to the pytorch norm, but that means all old model trained with old Flux norm will be affected.

I think we should mark this as breaking, and tag a breaking release