@CarloLucibello 2023-04-26
I think it is worth separating the speedup of the PR from the breaking change, since it may take some time to have a breaking release