@CarloLucibello 2023-04-27
It would be great if we could make the eps change not affect deserialized old models, I don't see how to do it though. I tried to think along the line of adding to the types and extra param acting as a flag, e.g. `struct LayerNorm{VERSION} ..` but this would break deserialization.